requirements.txt
ertifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.28.0
Flask==1.1.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
psycopg2==2.8.6
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.10
Werkzeug==1.0.1
